Insert into: added test of negative integer, unsigned integer, object id and object reference using highest bit in 64bits.

+++ src/Samples/QueryProcessingTest/WebVisitDataModel.cs
@@ -24,7 +24,7 @@ namespace QueryProcessingTest {
     // A Visit consists of Webpage Impressions
     [Database]
     public class Visit {
-        public int Id;
+        public ulong Id;
         public Company Company;
         public DateTime Start;
         public DateTime End;
